pyproject.toml
[tool.poetry]
name = "backend-boilerplate"
version = "0.1.0"
description = "Бойлерплейт бэкенда: FastAPI + Poetry + MongoDB (в облаке)"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pymongo = "^4.10.0"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
python-dotenv = "^1.0.0"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
httpx = "^0.28.1"
python-multipart = "^0.0.22"
email-validator = "^2.3.0"
boto3 = "^1.42.43"
python-slugify = "^8.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
